abstract out computTraceOptions
